Move `chia/types/eligible_coin_spends.py` to `chia/full_node`.
