+++ chia/_tests/core/mempool/test_mempool_manager.py
@@ -30,6 +30,13 @@ from chia._tests.util.misc import Marks, datacases, invariant_check_mempool
 from chia._tests.util.setup_nodes import OldSimulatorsAndWallets, setup_simulators_and_wallets
 from chia.consensus.condition_costs import ConditionCost
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
+from chia.full_node.eligible_coin_spends import (
+    DedupCoinSpend,
+    EligibilityAndAdditions,
+    IdenticalSpendDedup,
+    SkipDedup,
+    run_for_cost,
+)
 from chia.full_node.mempool import MAX_SKIPPED_ITEMS, PRIORITY_TX_THRESHOLD
 from chia.full_node.mempool_check_conditions import mempool_check_time_locks
 from chia.full_node.mempool_manager import (
@@ -60,13 +67,6 @@ from chia.types.coin_record import CoinRecord
 from chia.types.coin_spend import make_spend
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
-from chia.types.eligible_coin_spends import (
-    DedupCoinSpend,
-    EligibilityAndAdditions,
-    IdenticalSpendDedup,
-    SkipDedup,
-    run_for_cost,
-)
 from chia.types.mempool_inclusion_status import MempoolInclusionStatus
 from chia.types.mempool_item import BundleCoinSpend, MempoolItem, UnspentLineageInfo
 from chia.types.peer_info import PeerInfo
